+++ mod/CHANGES.md
@@ -118,6 +118,7 @@ While there are thousands of rooms in the game, many players have already seen t
 * The "Would you like to do a Victory Lap!?" popup no longer appears after defeating The Lamb.
 * All pills can now be used to cancel pedestal pickup animations.
 * The door to Hush is now automatically opened.
+* The devil statue will be faded if there is an item pedestal hiding behind it.
 
 ### Cutscene & Animation Removal